ADD Extract the Keystone Mock to a utility module
