+++ test/unit/validate_user_action_test.js
@@ -33,19 +33,6 @@ var serverMocks = require('../tools/serverMocks'),
     should = require('should'),
     request = require('request');
 
-function mockKeystone(req, res) {
-    if (req.path === '/v3/auth/tokens' && req.method === 'POST') {
-        res.setHeader('X-Subject-Token', '092016b75474ea6b492e29fb69d23029');
-        res.json(201, utils.readExampleFile('./test/keystoneResponses/authorize.json'));
-    } else if (req.path === '/v3/auth/tokens' && req.method === 'GET') {
-        res.json(200, utils.readExampleFile('./test/keystoneResponses/getUser.json'));
-    } else if (req.path === '/v3/projects' && req.method === 'GET') {
-        res.json(200, utils.readExampleFile('./test/keystoneResponses/getProjects.json'));
-    } else {
-        res.json(200, utils.readExampleFile('./test/keystoneResponses/rolesOfUser.json'));
-    }
-}
-
 function mockIdm(req, res) {
     if (req.path === '/user') {
         res.json(200, utils.readExampleFile('./test/authorizationResponses/rolesOfUser.json'));
@@ -83,7 +70,7 @@ describe('Validate action with Access Control', function() {
                 authenticationResponse: './test/keystoneResponses/authorize.json',
                 headers: [
                 ],
-                authMock: mockKeystone
+                authMock: serverMocks.mockKeystone
             }
         ];
